extract aion bootstrapping into a seaparate class based on consensus
