+++ modMcf/src/org/aion/mcf/config/CfgConsensus.java
@@ -27,6 +27,19 @@ import javax.xml.stream.XMLStreamReader;
 
 public abstract class CfgConsensus {
 
+    private ConsensusType consensusType;
+
+    public CfgConsensus(ConsensusType consensusType) {
+        this.consensusType = consensusType;
+    }
+
     public abstract void fromXML(final XMLStreamReader sr) throws XMLStreamException;
 
+    public ConsensusType getConsensusType() {
+        return consensusType;
+    }
+
+    public enum ConsensusType {
+        POW
+    }
 }
